project-nongye-tudouji/software/experimental/requirements.txt
# 农业巡检机器人实验版本 - 依赖包列表
# 安装命令: pip install -r requirements.txt

# 基础依赖
RPi.GPIO>=0.7.1
picamera>=0.13
opencv-python>=4.8.0
numpy>=1.21.0
Flask>=2.3.0

# 图像处理
Pillow>=9.0.0

# 数据处理
pandas>=1.3.0

# 配置文件处理
pyyaml>=6.0

# 系统监控
psutil>=5.8.0

# 串口通信（可选）
pyserial>=3.5

# 网络请求
requests>=2.28.0

# 时间处理
python-dateutil>=2.8.0

# 日志处理
colorlog>=6.0.0

# 数学计算
scipy>=1.7.0